+++ packages/hooks/typings/useCurrentRegion.d.ts
@@ -1,0 +1,8 @@
+type CurrentRegionType = {
+    code: string;
+    value: string;
+};
+
+declare function useCurrentRegion(): [CurrentRegionType | undefined, boolean];
+
+export default useCurrentRegion;