+++ packages/hooks/typings/useTimeout.d.ts
@@ -1,0 +1,3 @@
+declare function useTimeout(ms?: number | 0): boolean;
+
+export default useTimeout;