+++ packages/hooks/useEffectAsync.d.ts
@@ -1,3 +1,0 @@
-declare const useEffectAsync: Function;
-
-export default useEffectAsync;