+++ packages/feedback/CHANGELOG.md
@@ -3,6 +3,54 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [4.1.14](https://github.com/Availity/availity-react/compare/@availity/feedback@4.1.13...@availity/feedback@4.1.14) (2019-07-23)
+
+**Note:** Version bump only for package @availity/feedback
+
+
+
+
+
+## [4.1.13](https://github.com/Availity/availity-react/compare/@availity/feedback@4.1.12...@availity/feedback@4.1.13) (2019-07-19)
+
+**Note:** Version bump only for package @availity/feedback
+
+
+
+
+
+## [4.1.12](https://github.com/Availity/availity-react/compare/@availity/feedback@4.1.11...@availity/feedback@4.1.12) (2019-07-11)
+
+**Note:** Version bump only for package @availity/feedback
+
+
+
+
+
+## [4.1.11](https://github.com/Availity/availity-react/compare/@availity/feedback@4.1.10...@availity/feedback@4.1.11) (2019-06-24)
+
+**Note:** Version bump only for package @availity/feedback
+
+
+
+
+
+## [4.1.10](https://github.com/Availity/availity-react/compare/@availity/feedback@4.1.9...@availity/feedback@4.1.10) (2019-06-20)
+
+**Note:** Version bump only for package @availity/feedback
+
+
+
+
+
+## [4.1.9](https://github.com/Availity/availity-react/compare/@availity/feedback@4.1.8...@availity/feedback@4.1.9) (2019-06-17)
+
+**Note:** Version bump only for package @availity/feedback
+
+
+
+
+
 ## [4.1.8](https://github.com/Availity/availity-react/compare/@availity/feedback@4.1.7...@availity/feedback@4.1.8) (2019-06-03)
 
 **Note:** Version bump only for package @availity/feedback