+++ packages/link/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.2.1](https://github.com/Availity/availity-react/compare/@availity/link@1.2.0...@availity/link@1.2.1) (2019-06-24)
+
+**Note:** Version bump only for package @availity/link
+
+
+
+
+
 # [1.2.0](https://github.com/Availity/availity-react/compare/@availity/link@1.1.2...@availity/link@1.2.0) (2019-06-11)