+++ packages/progress/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.1.1](https://github.com/Availity/availity-react/compare/@availity/progress@1.1.0...@availity/progress@1.1.1) (2019-06-24)
+
+**Note:** Version bump only for package @availity/progress
+
+
+
+
+
 # 1.1.0 (2019-06-06)