+++ packages/upload/CHANGELOG.md
@@ -3,6 +3,38 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [2.0.4](https://github.com/Availity/availity-react/compare/@availity/upload@2.0.3...@availity/upload@2.0.4) (2019-07-19)
+
+**Note:** Version bump only for package @availity/upload
+
+
+
+
+
+## [2.0.3](https://github.com/Availity/availity-react/compare/@availity/upload@2.0.2...@availity/upload@2.0.3) (2019-07-02)
+
+**Note:** Version bump only for package @availity/upload
+
+
+
+
+
+## [2.0.2](https://github.com/Availity/availity-react/compare/@availity/upload@2.0.1...@availity/upload@2.0.2) (2019-06-24)
+
+**Note:** Version bump only for package @availity/upload
+
+
+
+
+
+## [2.0.1](https://github.com/Availity/availity-react/compare/@availity/upload@2.0.0...@availity/upload@2.0.1) (2019-06-20)
+
+**Note:** Version bump only for package @availity/upload
+
+
+
+
+
 # [2.0.0](https://github.com/Availity/availity-react/compare/@availity/upload@1.8.2...@availity/upload@2.0.0) (2019-06-03)